refactor: move book member queries to schema
